機能を分割し、コードのモジュール化を実施
`level_editor.py` から多くのコードを削除し、`stretch_vertex.py`、`create_ico_sphere.py`、`export_scene.py`、`collider_system.py` という新しいファイルに機能を分割しました。これにより、可読性と保守性が向上しました。

`__init__.py` では新しいオペレーターをインポートし、アドオンのクラスとして登録。アドオンの有効化および無効化時のコールバック関数も追加しました。

`collider_system.py` にはコライダーの描画と管理に関するクラスが追加され、`create_ico_sphere.py` ではICO球を生成するオペレーターが実装されました。

`export_scene.py` ではシーン情報をJSON形式でエクスポートする機能が強化され、`stretch_vertex.py` では特定のオブジェクトの頂点を伸ばすオペレーターが定義されています。
