--- Tool/level_editor.py
@@ -1,458 +1,0 @@
-import bpy
-import math
-import bpy_extras
-import gpu
-import gpu_extras.batch
-import copy
-import mathutils
-import json
-
-
-bl_info = {
-    "name": "レベルエディタ",
-    "author": "Taiki Ozawa",
-    "version": (1 ,0),
-    "blender": (3 ,3 ,1),
-    "location": "",
-    "description": "レベルエディタ",
-    "warning": "",		
-    "wiki_url": "",
-    "tracker_url": "",
-    "category": "Object"
-}
-
-class MYADDON_OT_stretch_vertex(bpy.types.Operator):
-    bl_idname = "myaddon.myaddon_t_stretch_vertex"
-    bl_label = "頂点を伸ばす"
-    bl_description = "頂点座標を引っ張ってのばす"
-
-    #Redo,Undo可能オプション
-    bl_option = {'REGISTER' , 'UNDO'}
-
-    def execute(self,context):
-        bpy.data.objects["Cube"].data.vertices[0].co.x += 1.0
-        print("頂点をのばしました")
-
-        return {'FINISHED'}
-
-
-class MYADDON_OT_create_ico_sphere(bpy.types.Operator):
-    bl_idname = "myaddon.myaddon_create_ico_sphere"
-    bl_label = "ICO球生成"
-    bl_description = "ICO球を生成します"
-
-    #Redo,Undo可能オプション
-    bl_option = {'REGISTER' , 'UNDO'}
-
-    def execute(self,context):
-        bpy.ops.mesh.primitive_ico_sphere_add()
-        print("ICO球を生成しました")
-
-        return {'FINISHED'} 
-    
-class MYADDON_OT_export_scene(bpy.types.Operator,bpy_extras.io_utils.ExportHelper):
-    bl_idname = "myaddon.myaddon_ot_export_scene"
-    bl_label = "シーン出力"
-    bl_description = "シーン情報をexportします"
-    #出力するファイルの拡張子
-    filename_ext=".json"
-
-    def execute(self,context):
-        print("シーン情報を出力します")
-        print(bpy.context.scene.objects)
-        
-        self.export_json()
-
-        print("シーン情報を出力しました")
-        self.report({'INFO'},"シーン情報を出力しました")
-
-        return {'FINISHED'}
-
-    def export(self):
-
-        print("シーン情報を出力開始．．．%r" % self.filepath)
-        
-        #ファイルをテキスト形式で書き出しようにオープン
-        #スコープを抜けると自動でクローズされる
-        with open(self.filepath,"wt") as file:
-            #ファイルに文字列を書き込む
-            file.write("SCENE\n")
-
-            for object in bpy.context.scene.objects:
-                if(object.parent):
-                    continue
-            
-                self.parse_scene_recursive(file,object,0)
-
-    def export_json(self):
-        """シーン情報をjson形式で書き出し"""
-
-        #保存する情報をまとめるdict
-        json_object_root = dict()
-
-        #ノード名
-        json_object_root["name"] = "scene"
-        #オブジェクトリストを作成
-        json_object_root["objects"] = list()
-
-        #シーン内の全てのオブジェクトを走査
-        for object in bpy.context.scene.objects:
-            #親オブジェクトはスキップ
-            if(object.parent):
-                continue
-            
-            #シーン直下のオブジェクトをルートノードとして再帰で走査
-            self.parse_scene_recursive_json(json_object_root["objects"],object,0)
-            
-
-
-        json_text = json.dumps(json_object_root,ensure_ascii=False,cls=json.JSONEncoder,indent =4)
-        #コンソールに表示
-        print(json_text)
-
-        #ファイルをテキスト形式で書き出し用にオープン
-        #スコープを抜けると自動でクローズされる
-        with open(self.filepath,"wt",encoding="utf-8") as file:
-            #ファイルに文字列を書き込む
-            file.write(json_text)
-            
-
-    def write_and_print(self,file,str):
-        print(str)
-
-        file.write(str)
-        file.write("\n")
-
-    def parse_scene_recursive(self, file, object, level):
-        """シーン解析用再帰関数"""
-
-        indent = ''
-        for i in range(level):
-            indent += "\t"
-
-        # オブジェクト名書き込み
-        self.write_and_print(file,indent + object.type )
-        trans, rot, scale = object.matrix_local. decompose( )
-        #回転を Quternion から Euler (3軸での回転角)に変換
-        rot = rot. to_euler( )
-        #ラジアンから度数法に変換
-        rot.x = math.degrees(rot.x)
-        rot.y = math.degrees(rot.y)
-        rot.z = math.degrees(rot.z)
-        #トランスフォーム情報を表示
-        self.write_and_print(file, indent + "T %f %f %f" % (trans.x, trans.y, trans. z) )
-        self.write_and_print(file, indent + "R %f %f %f" % (rot.x, rot.y, rot. z) )
-        self.write_and_print(file, indent + "S %f %f %f" % (scale.x, scale.y, scale.z) )
-
-        #カスタムプロパティの書き込み
-        if "file_name" in object:
-            self.write_and_print(file,indent + "N %s" % object["file_name"])
-        
-        # Colliderの書き出し
-        if "collider" in object:
-            self.write_and_print(file,indent + "C %s" % object["collider"])
-            temp_str = indent + "CC %f %f %f"
-            temp_str %= (object["collider_center"][0], object["collider_center"][1], object["collider_center"][2])
-            self.write_and_print(file,temp_str)
-            temp_str = indent + "CS %f %f %f"
-            temp_str %= (object["collider_size"][0], object["collider_size"][1], object["collider_size"][2])
-            self.write_and_print(file,temp_str)
-
-        self.write_and_print(file,indent + "END")
-        self.write_and_print(file, '')
-
-        for child in object.children:
-            self.parse_scene_recursive(file ,child ,level + 1)
-
-
-    def parse_scene_recursive_json(self, data_parent, object, level):
-        #シーンのオブジェクト一個分のjsonデータを作成
-        json_object = dict()
-        #objectの種類
-        json_object["type"] = object.type
-        #オブジェクト名
-        json_object["name"] = object.name
-
-        #オブジェクトのローカルトランスフォームをから
-        #平行移動 回転 スケールを取得
-        trans,rot,scale = object.matrix_local.decompose()
-        #回転を Quternion から Euler (3軸での回転角)に変換
-        rot = rot.to_euler()
-        #ラジアンから度数法に変換
-        rot.x = math.degrees(rot.x)
-        rot.y = math.degrees(rot.y)
-        rot.z = math.degrees(rot.z)
-
-        #トランスフォーム情報をディクショナリに登録
-        transform = dict()
-        transform["transform"] = (trans.x,trans.y,trans.z)
-        transform["rotation"] = (rot.x,rot.y,rot.z)
-        transform["scale"] = (scale.x,scale.y,scale.z)
-
-        #トランスフォーム情報をjsonオブジェクトに登録
-        json_object["transform"] = transform
-
-        #カスタムプロパティの書き込み
-        if "file_name" in object:
-            #カスタムプロパティが存在する場合は登録
-            json_object["file_name"] = object["file_name"]
-
-        if "collider" in object:
-            #カスタムプロパティが存在する場合は登録
-            collider = dict()
-            collider["type"] = object["collider"]
-            collider["center"] = object["collider_center"].to_list()
-            collider["size"] = object["collider_size"].to_list()
-            json_object["collider"] = collider
-    
-
-
-        #一個分のjsonオブジェクトを親オブジェクトに登録
-        data_parent.append(json_object)
-
-        #直接子供のリストを走査
-        if len(object.children) > 0:
-            #子ノードリストを作成
-            json_object["children"] = list()
-
-            #子供のリストを走査
-            for child in object.children:
-                #子供のjsonデータを作成
-                self.parse_scene_recursive_json(json_object["children"],child,level + 1)
-
-
-
-class MYADDON_OT_add_filename(bpy.types.Operator):
-    bl_idname = "myaddon.myaddon_add_filename"
-    bl_label = "FileName 追加"
-    bl_description = "['file_name']カスタムプロパティを追加します"
-    bl_option = {'REGISTER' , 'UNDO'}
-
-    def execute(self,context):
-        #['file_name']カスタムプロパティを追加
-        context.object["file_name"] = ""
-
-        return {'FINISHED'}            
-    
-
-#トップバーの拡張メニュー
-class TOPBAR_MT_my_menu(bpy.types.Menu):
-    #Blenderがクラスを識別するための固有の文字列
-    bl_idname= "TOPBAR_MT_my_menu"
-    #メニューのラベルとして表示される文字列
-    bl_label = "MyMenu"
-    #著者表示用の文字列
-    bl_description = "拡張メニュー by" + bl_info["author"]
-
-    def draw(self,context):
-        #self : 呼び出し元のクラスインスタンス c++のthisポインタ
-        #contet : カーソルを合わせたときのポップアップのカスタマイズなどに使用
-
-        #self.layout.operator("wm.url_open_preset",text = "Manual", icon ='HELP')
-        #self.layout.separator()
-        #self.layout.operator("wm.url_open_preset",text = "Manual", icon ='HELP')
-        self.layout.operator(MYADDON_OT_stretch_vertex.bl_idname,text=MYADDON_OT_stretch_vertex.bl_label)
-        self.layout.operator(MYADDON_OT_create_ico_sphere.bl_idname,text=MYADDON_OT_create_ico_sphere.bl_label)
-        self.layout.operator(MYADDON_OT_export_scene.bl_idname,text=MYADDON_OT_export_scene.bl_label)
-
-
-    def submenu(self,context):
-        #self : 呼び出し元のクラスインスタンス c++のthisポインタ
-        #contet : カーソルを合わせたときのポップアップのカスタマイズなどに使用
-
-        self.layout.menu(TOPBAR_MT_my_menu.bl_idname)
-
-
-class OBJECT_PT_file_name(bpy.types.Panel):
-    """ オブジェクトのファイルネームパネル """
-    bl_idname = "OBJECT_PT_file_name"
-    bl_label = "FIleName"
-    bl_space_type = "PROPERTIES"
-    bl_region_type = "WINDOW"
-    bl_context = "object"
-
-    #サブメニューの描画
-    def draw(self,context):
-
-        if "file_name" in context.object:
-            #カスタムプロパティが存在する場合は表示
-            self.layout.prop(context.object, '["file_name"]', text = self.bl_label)
-        else:
-            #カスタムプロパティが存在しない場合は追加ボタンを表示
-            self.layout.operator(MYADDON_OT_add_filename.bl_idname)
-
-        #パネルの項目を追加
-        #self.layout.operator(MYADDON_OT_stretch_vertex.bl_idname, text=MYADDON_OT_stretch_vertex.bl_label)
-        #self.layout.operator(MYADDON_OT_create_ico_sphere.bl_idname, text=MYADDON_OT_create_ico_sphere.bl_label)
-        #self.layout.operator(MYADDON_OT_export_scene.bl_idname, text=MYADDON_OT_export_scene.bl_label)
-
-
-class DrawCollider:
-
-    #描画ハンドル
-    handler = None
-
-    def draw_collider():
-        #頂点データ
-        vertices ={"pos":[]}
-        #インデックスデータ
-        indices = []
-
-        offsets=[
-            [-0.5,-0.5,-0.5],#左下前
-            [+0.5,-0.5,-0.5],#右下前
-            [-0.5,+0.5,-0.5],#左上前
-            [+0.5,+0.5,-0.5],#右上前
-            [-0.5,-0.5,+0.5],#左下奥
-            [+0.5,-0.5,+0.5],#右下奥
-            [-0.5,+0.5,+0.5],#左上奥
-            [+0.5,+0.5,+0.5]#右上奥            
-        ]
-        size = [2,2,2]
-
-        for object in bpy.context.scene.objects:
-
-            #コライダープロパティがなければスキップ
-            if "collider" not in object:
-                continue
-
-            center = mathutils.Vector((0,0,0))
-            size = mathutils.Vector((2,2,2))
-
-            #プロパティからコライダーの値を取得
-            center[0]= object["collider_center"][0]
-            center[1]= object["collider_center"][1]
-            center[2]= object["collider_center"][2]
-            size[0]= object["collider_size"][0]
-            size[1]= object["collider_size"][1]
-            size[2]= object["collider_size"][2]
-
-            #追加前の頂点数
-            start = len(vertices["pos"])
-
-            #boxの8頂点分回す
-            for offset in offsets:
-                #オブジェクトの中心座標をコピー
-                pos = copy.copy(center)
-                #中心座標を基準に各頂点ごとにずらす
-                pos[0] += offset[0] * size[0]
-                pos[1] += offset[1] * size[1]
-                pos[2] += offset[2] * size[2]
-
-                #オブジェクトのローカル座標をワールド座標に変換
-                pos = object.matrix_world @ pos
-
-                #頂点データに追加
-                vertices["pos"].append(pos)
-
-                indices.append([start+0,start+1])
-                indices.append([start+2,start+3])
-                indices.append([start+0,start+2])
-                indices.append([start+1,start+3])
-
-                indices.append([start+4,start+5])
-                indices.append([start+6,start+7])
-                indices.append([start+4,start+6])
-                indices.append([start+5,start+7])
-
-                indices.append([start+0,start+4])
-                indices.append([start+1,start+5])
-                indices.append([start+2,start+6])
-                indices.append([start+3,start+7])
-
-
-        #ビルトインのシェーダーを使用
-        shader = gpu.shader.from_builtin("UNIFORM_COLOR")
-        #バッチを作成
-        batch = gpu_extras.batch.batch_for_shader(shader, 'LINES', vertices, indices=indices)
-
-        #シェーダーのパラメータを設定
-        color = {0.5, 1.0, 1.0, 1.0}
-        shader.bind()
-        shader.uniform_float("color", color)
-        #バッチを描画
-        batch.draw(shader)
-
-
-class MYADDON_OT_add_collider(bpy.types.Operator):
-    bl_idname = "myaddon.myaddon_add_collider"
-    bl_label = "Collider 追加"
-    bl_description = "[Collider]カスタムプロパティを追加します"
-    bl_option = {'REGISTER' , 'UNDO'}
-
-    #Redo,Undo可能オプション
-    bl_option = {'REGISTER' , 'UNDO'}
-
-    def execute(self,context):
-        
-        #['Collider']カスタムプロパティを追加
-        context.object["collider"] = "BOX"
-        context.object["collider_center"] = mathutils.Vector((0,0,0))
-        context.object["collider_size"] = mathutils.Vector((2,2,2))
-
-        return {'FINISHED'}
-
-class OBJECT_PT_collider(bpy.types.Panel):
-    """ オブジェクトのColliderパネル """
-    bl_idname = "OBJECT_PT_collider"
-    bl_label = "Collider"
-    bl_space_type = "PROPERTIES"
-    bl_region_type = "WINDOW"
-    bl_context = "object"
-
-    #サブメニューの描画
-    def draw(self,context):
-
-        if "collider" in context.object:
-            #カスタムプロパティが存在する場合は表示
-            self.layout.prop(context.object, '["collider"]', text = self.bl_label)
-            self.layout.prop(context.object, '["collider_center"]', text = "Center")
-            self.layout.prop(context.object, '["collider_size"]', text = "Size")
-
-        else:
-            #カスタムプロパティが存在しない場合は追加ボタンを表示
-            self.layout.operator(MYADDON_OT_add_collider.bl_idname)
-
-        #パネルの項目を追加
-        #self.layout.operator(MYADDON_OT_stretch_vertex.bl_idname, text=MYADDON_OT_stretch_vertex.bl_label)
-        #self.layout.operator(MYADDON_OT_create_ico_sphere.bl_idname, text=MYADDON_OT_create_ico_sphere.bl_label)
-        #self.layout.operator(MYADDON_OT_export_scene.bl_idname, text=MYADDON_OT_export_scene.bl_label)
-
-classes = {
-    MYADDON_OT_stretch_vertex,
-    MYADDON_OT_create_ico_sphere,
-    MYADDON_OT_export_scene,
-    MYADDON_OT_add_filename,
-    TOPBAR_MT_my_menu,
-    OBJECT_PT_file_name,
-    MYADDON_OT_add_collider,
-    OBJECT_PT_collider,
-}
-
-
-#アドオン有効化時のコールバック
-def register():
-    for cls in classes:
-        bpy.utils.register_class(cls)
-
-    bpy.types.TOPBAR_MT_editor_menus.append(TOPBAR_MT_my_menu.submenu)
-    # 3Dビューに描画関数を追加
-    DrawCollider.handler = bpy.types.SpaceView3D.draw_handler_add(DrawCollider.draw_collider, (), 'WINDOW', 'POST_VIEW')
-    print("レベルエディタが有効化されました")
-
-#アドオン無効化時のコールバック
-def unregister():
-    bpy.types.TOPBAR_MT_editor_menus.remove(TOPBAR_MT_my_menu.submenu)
-    # 3Dビューから描画関数を削除
-    bpy.types.SpaceView3D.draw_handler_remove(DrawCollider.handler, 'WINDOW')
-
-    for cls in classes:
-        bpy.utils.unregister_class(cls)
-        
-    print("レベルエディタが無効化されました。")
-
-
-
-#テスト実行用コード
-if __name__ == "__main__":
-    register()
