+++ Tool/level_editor/stretch_vertex.py
@@ -1,0 +1,15 @@
+import bpy
+
+class MYADDON_OT_stretch_vertex(bpy.types.Operator):
+    bl_idname = "myaddon.myaddon_t_stretch_vertex"
+    bl_label = "頂点を伸ばす"
+    bl_description = "頂点座標を引っ張ってのばす"
+
+    #Redo,Undo可能オプション
+    bl_option = {'REGISTER' , 'UNDO'}
+
+    def execute(self,context):
+        bpy.data.objects["Cube"].data.vertices[0].co.x += 1.0
+        print("頂点をのばしました")
+
+        return {'FINISHED'}
